Move submit handling to a separate file
